feat: findidview navigation연결 및 loginview 화면 아이콘 추가
