+++ pennyway-client-iOS/pennyway-client-iOS/View/LoginView/AdditionalOptionView.swift
@@ -10,7 +10,14 @@ struct AdditionalOptionView: View {
             .font(.pretendard(.medium, size: 9))
             .platformTextColor(color: Color("Gray04"))
 
-            NavigationLink(destination: FindIDView()) {
+            Image("icon_line_gray")
+                .frame(width: 0, height: 9)
+                .overlay(
+                    Rectangle()
+                        .stroke(Color("Gray02"), lineWidth: 0.69506)
+                )
+
+            NavigationLink(destination: FindIDFormView()) {
                 Text("아이디 찾기")
             }
             .font(.pretendard(.medium, size: 9))
